requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==4.2.2
altgraph==0.17.4
annotated-types==0.7.0
ansi2html==1.9.1
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
blis==0.7.11
boto3==1.39.13
botocore==1.39.13
cachetools==5.3.2
captum==0.6.0
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chart-studio==1.1.0
click==8.2.1
cloudpathlib==0.19.0
cmdstanpy==1.2.4
colorama==0.4.6
comm==0.2.1
confection==0.1.5
contourpy==1.2.0
convertdate==2.4.0
cryptography==43.0.3
cycler==0.11.0
cymem==2.0.8
Cython==3.0.8
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_renderer==1.9.1
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
Django==4.2
dnspython==2.7.0
download==0.3.5
ecdsa==0.19.1
email_validator==2.2.0
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.2.0
executing==2.0.1
Faker==25.5.0
fastapi==0.115.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fastjsonschema==2.19.1
favicon==0.7.0
feedparser==6.0.11
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.3.25
fonttools==4.28.3
fpdf==1.7.2
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2023.12.2
future==1.0.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
h5py==3.11.0
hijri-converter==2.3.1
holidays==0.57
htbuilder==0.6.2
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.5
idna==3.6
imageio==2.36.0
importlib-metadata==7.0.1
importlib_resources==6.4.5
ipykernel==6.28.0
ipython==8.19.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.5.0
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-dash==0.4.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.0
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kafka-python==2.2.11
kaggle==1.6.17
keras==3.7.0
kiwisolver==1.3.2
korean-lunar-calendar==0.3.1
langchain==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langcodes==3.4.0
langsmith==0.1.117
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
lightning-utilities==0.10.1
llama-cloud==0.0.17
llama-index==0.11.8
llama-index-agent-openai==0.3.1
llama-index-cli==0.3.1
llama-index-core==0.11.8
llama-index-embeddings-openai==0.2.4
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.3
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.1
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.3
llvmlite==0.43.0
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
marshmallow==3.22.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.4
multitasking==0.0.11
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
namex==0.0.8
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.3
neuralprophet==0.6.2
nltk==3.9.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
openai-whisper==20240930
opencv-python==4.10.0.82
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
overrides==7.4.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
patsy==1.0.1
peewee==3.17.6
pefile==2023.2.7
pillow==10.3.0
platformdirs==4.1.0
plotly==5.18.0
plotly-resampler==0.8.3.2
pooch==1.8.2
preshed==3.0.9
prettytable==3.10.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
prophet==1.1.5
proto-plus==1.25.0
protobuf==3.20.3
psutil==5.9.7
psycopg2==2.9.9
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==16.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.9.1
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.23.3
pydeck==0.8.1b0
pydub==0.25.1
pyecharts==2.0.5
pygame==2.6.1
Pygments==2.17.2
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyJWT==2.8.0
pymdown-extensions==10.8.1
PyMeeus==0.5.12
PyMuPDF==1.25.0
PyMySQL==1.1.0
pyodbc==5.2.0
pyparsing==3.0.6
pypdf==4.3.1
PyPDF4==1.27.0
pyspark==3.5.5
pytesseract==0.3.13
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
python-json-logger==2.0.7
python-math==0.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytorch-lightning==1.9.5
pytube==15.0.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2024.7.24
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rich-toolkit==0.14.8
rignore==0.6.4
rpds-py==0.16.2
rsa==4.9
s3transfer==0.13.1
safetensors==0.4.5
scikit-learn==1.4.1.post1
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.2
sentry-sdk==2.33.2
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.5.0.post1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
sqlparse==0.4.4
srsly==2.4.8
st-annotated-text==4.0.1
st-pages==0.5.0
st-star-rating==0.0.5
st-theme==1.2.3
stack-data==0.6.3
stanio==0.5.1
starlette==0.38.6
statsmodels==0.14.4
streamlit-aggrid==1.0.5
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-echarts==0.4.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-image-select==0.6.0
streamlit-keyup==0.2.4
streamlit-lightweight-charts==0.7.20
streamlit-modal==0.1.2
streamlit-option-menu==0.3.12
streamlit-star-ratings==0.0.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
striprtf==0.0.26
sympy==1.14.0
tenacity==8.2.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.4.0
terminado==0.18.0
text-unidecode==1.3
tf_keras==2.18.0
thinc==8.2.5
threadpoolctl==3.3.0
tiktoken==0.7.0
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.7.1
torchaudio==2.7.1
torchmetrics==0.11.4
tornado==6.4
tqdm==4.66.1
trace-updater==0.0.9.1
traitlets==5.14.1
transformers==4.47.0
typer==0.16.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.35.0
validators==0.28.3
wasabi==1.1.3
watchdog==4.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
whisper==1.1.10
widgetsnbextension==4.0.9
winshell==0.6
wrapt==1.16.0
WTForms==3.2.1
xgboost==3.0.2
yarl==1.9.4
yfinance==0.2.40
zipp==3.17.0
